feat: preserve aspect ratio when generating thumbnails
The thumbnails will still be 140px tall, only we're no longer forcing the width to be 80px. The width will be whatever it needs to be to preserve the aspect ratio. The frontend code which displays thumbnails makes sure that even if some really thin aspect ratio is used that it won't break the UI.